+++ demo/api/src/app.module.ts
@@ -61,7 +61,6 @@ export class AppModule {
                         debug: config.debug,
                         playground: config.debug,
                         autoSchemaFile: "schema.gql",
-                        sortSchema: true,
                         context: ({ req }: { req: Request }) => ({ ...req }),
                         cors: {
                             credentials: true,