+++ packages/api/cms-api/src/builds/changes-checker.interceptor.ts
@@ -3,12 +3,17 @@ import { Reflector } from "@nestjs/core";
 import { GqlExecutionContext } from "@nestjs/graphql";
 import { Observable } from "rxjs";
 
+import { ContentScopeService } from "../content-scope/content-scope.service";
 import { BuildsService } from "./builds.service";
 import { SKIP_BUILD_METADATA_KEY } from "./skip-build.decorator";
 
 @Injectable()
-export class SkipBuildInterceptor implements NestInterceptor {
-    constructor(private reflector: Reflector, private readonly buildsService: BuildsService) {}
+export class ChangesCheckerInterceptor implements NestInterceptor {
+    constructor(
+        private reflector: Reflector,
+        private readonly buildsService: BuildsService,
+        private readonly contentScopeService: ContentScopeService,
+    ) {}
 
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     async intercept(context: ExecutionContext, next: CallHandler): Promise<Observable<any>> {
@@ -20,7 +25,9 @@ export class SkipBuildInterceptor implements NestInterceptor {
                     this.reflector.get<string[]>(SKIP_BUILD_METADATA_KEY, context.getClass());
 
                 if (!skipBuild) {
-                    await this.buildsService.setChangesSinceLastBuild();
+                    const scope = await this.contentScopeService.inferScopeFromExecutionContext(context);
+
+                    await this.buildsService.setChangesSinceLastBuild(scope);
                 }
             }
         }