+++ packages/api/cms-api/src/builds/dto/build.object.ts
@@ -1,6 +1,7 @@
 import { Field, ID, ObjectType } from "@nestjs/graphql";
 
 import { JobStatus } from "../../kubernetes/job-status.enum";
+import { LABEL_ANNOTATION } from "../builds.constants";
 
 @ObjectType()
 export class Build {
@@ -12,6 +13,9 @@ export class Build {
 
     @Field({ nullable: true })
     name?: string;
+
+    @Field({ nullable: true, description: `Human readable label provided by ${LABEL_ANNOTATION} annotation. Use name as fallback if not present` })
+    label?: string;
 
     @Field({ nullable: true })
     trigger?: string;